requirements.txt
Flask
Flask-MySQL
pandas
matplotlib
